pyproject.toml
[project]
name = "mf07-language"
version = "1.0.7"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "pyyaml",
]
